Extending the feature introduced in the previous revision, this one adds automatic tracing of the command lines used to compile different classes (C, Fortran, and executables). So now it is not necessary to track makefiles or options used in their invocation - only the final command line (name of compiler and flags) matters. This information is kept in .*opts files in both seq and mpi folders.
Another change is that common parts of specific Makefiles was moved to a separate file common.mk that is included from each of the former. This common file is responsible for processing both dependencies and .*opts files.

Also, most of variable assignments in makefiles were changed to immediate ":=", thus it is easier to keep track of their values, especially in the main Makefile.