'-int igt_so' was implemented. That is Integrated Green's tensor computed approximately using series expansion in kd up to second order. It reuses tabulated integrals and part of the code of SO. Basically, it is the same as SO, but with removed part of the code that depends on m. So 'igt_so' it is almost as fast as the default 'poi' and almost as accurate as 'igt' (and slightly more accurate than, say, 'igt 3'. Fixes issue 84.
CalcInterTerm is slightly optimized. The significant change is that SO formulation do not breaks when table size is not enough, but uses 'medium' formula instead of 'close' one. The same algorithm works for newly implemented 'igt_so'. Also debugging code to print values of G was added.
Function MultScal in cmplx.h renamed to vMultScal, and function vMultScalSelf added.
Version incremented to 1.1a2