Updated the default build platform to gcc v.4.x (tested with 4.5.2). Many small changes to remove new warnings. In particular, testing the version of gcc was added for some attributes. Fixes issue 122.
Added compiler options for Intel to handle C++ files, but they were not tested due to current incompatibility with OpenCL headers from CUDA 3.2.

Saving of used command lines for compilation to file (in common.mk) is now more robust. Should work fine with, e.g., paths in quotes. In principle, it may need further improvement to handle any special characters.

InitShape() in make_particle.c was modified not to use box_det_sh and size_det_sh, but instead to get equivalent information based on which variables are and are not initialized for a particular shape. This should make implementing new shapes a bit easier.

Testing for windows was updates to use not only _WIN32, but also _WIN64 - should make it more suitable for compilation on 64-bit Windows.

Added mentioning of MPICH2 or OpenMPI versions in output of 'adda -V', as well as version of MinGW environment (if present).

Version incremented to 1.1b1.