New order of headers in all source files is: - const.h (see below) - corresponding header (*.c->*.h) - project headers - 3rd party headers - system headers Order inside groups is alphabetical. This is more reliable for development (e.g. sefl-sufficiency of corresponding headers is tested automatically during compilation). The only out-of-way header is const.h, which is now included always in the first place. This is because it contains definition of macro __USE_MINGW_ANSI_STDIO, which tells MinGW64 to use c99-compliant printf-family functions. This macro should be defined before any of the system headers. Overall, this fixes issue with %z format specifiers for 64bit Windows.
Major bug in Makefile was fixed. It was using CPATH variable for communicating with child Makefiles. The side effect is that gcc was adding the project file to search path of system headers. Thus project headers were included instead system ones in certain cases (e.g. io.h). This was causing errors in compilation of C++ files, and warnings during compilation of ADDAmain.c on MinGW64. But more hidden effects could be possible. Overall, we should be more careful in variable naming inside the main Makefile (since they are exported).

Compiled win64 executables (both for ADDA and misc) are now collected in win64/ folder. The only problem is ocl executable - the compilation went OK, but the executable produces fatal errors when starts interacting with the GPU - needs further investigation.

Minor changes to definition of '-shape prism' in param.c and to svn properties of win32/.

%z printf format specifiers are removed from fft_setup.cpp, the functionality is imitated by casts to unsigned long. This saves a lot of hassle, since enforcing c99 standard on C++ files is somewhat confusing.