Implemented automatic counting of the OpenCL memory through wrappers for creating and releasing buffers. The good thing is that OpenCL allows easy querying of actual size of any memory object. Additionally detection of leaks of OpenCL memory is implemented.
Counted memory (both peak total and maximum for a single object) is given in the output (also operational with -prognosis) - they can be compared against corresponding maximum values for a particular device, which are also shown. However, the maximum memory for a single object, given by the device, doesn't seem to be a hard limit - at least, Nvidia cards work fine when ADDA exceeds this limit (but fits into the total memory).

Most of the memory required for standard matvec (Dmatrix, Xmatrix, etc.) is now not allocated (and some are even not defined) in the OpenCL version. This approximately halfs the CPU memory usage. So now memory usage of sequential mode is approximately equally divided between CPU and GPU in ocl mode. We also implemented by-hand counting of peak and current CPU memory, which is important since ocl version still needs a lot of temporary storage when Dmatrix is initialized (before it is copied to the GPU). So now the CPU memory reported, should accurately correspond to the peak memory. 

Moreover, the memory report at the end of AllocateEverything() in calculator.c does not ruin the value of 'memory' anymore, so it can be further updated if needed. For that format of AccumulateMax() in comm.c was changed.

Xmatrix was removed from vars.c/h and declared as extern in iterative.c in non-OpenCL mode (as a storage for arrays in WKB initial field). However, for OpenCL mode a complicated logic was added - these arrays either use storage of rvec,xvec,and Avecbuffer (if fits) or allocate extra memory. This should require no extra memory in most cases.

Mode of execution of OpenCL context was changed to 'in-order', hence several calls to clFinish(command_queue) could be removed from matvec.c.

OpenCL is significantly simplified by adding dynamically-allocating wrappers for clGet...Info functions. Also added dynamically-allocating version of sprintf to io.c/h. It is now called from make_particle.c and oclcore.c

blockTranspose in matvec.c is now called only in PARALLEL.

added MAXIMIZE macro to const.h

version incremented to 1.1b6

tests/2exec/comp2exec was modified to ignore new information, produced by adda_ocl, in ocl_seq mode.