Removed wrong comment about using xz-symmetry plane during integration over phi. It was present there since 2006 and could lead to incorrect results.
A correct way, as has been mentioned in the manual, is to specify min=0, max=180 (instead of -90 and 90), but even this requires proper handling of off-diagonal elements of the Mueller matrix (some of them no-more integrate to zero over the half-range). Therefore, it is better to keep this discussion (with all details) in the manual.