Implemented calculation of the scattered fields in the presence of surface (scattering directions both above and below). It is located in a separate function CalcFieldSurf(). There are, however, certain limitations for scattering below the surface when refractive index of the substrate is complex (not real). The main problem is that the scattered wave in such medium is inhomogeneous and p-polarization is determined by a complex unit vector (such that e.e=1). This complication is not yet accounted for in the main part of the code (that would be very complicated, especially in the case of some nontrivial scattering planes). So the calculated field itself is correct but it is then processed incorrectly further on (though correctness very much depends on a particular definition).
Also during calculation of Csca and g*Csca multipliers Re(msub) and Re(msub)^2 were added to |E|^2 for directions below the substrate. Thus those formula correspond to the total energy and momentum carried away by the scattered field for real msub. For complex msub it is still unclear, what should be the proper formulae (see comments in the code). Also the total momentum doesn't determine the total force even for non-absorbing substrate, since the substrate also absorbs momentum through reflection/refraction.

As a test, I've studied the convergence of Csca to the free-space result when msub->1. The difference in Csca seems to be O(sqrt(m-1)), but very large Jmax in integration over theta is required for proper analysis (up to 15-20). The latter is due to very sharp dependence of scattered intensity for theta near 90 degrees.

Other things:
- added functions vInit, cvInit, and vAlongZ to cmplx.h. The latter was moved from param.c
- function AsymParm() marked as unused.
- code in InitBeam() was slightly changed for surface, not to lose accuracy for near-z propagation.
- updsvnrev.sh was rewritten to avoid bashisms, so that it can be run under any shell (issue 172)