Added rigorous calculations of Sommerfeld integrals. Imported somnec.c from NEC2C, which computes the integrals themselves. The remaining code is in interaction.c - it creates a table of integrals (for different rho and z) and further indexes it to calculate the reflected Green's tensor. There is a lot that can be improved, but in sparse mode it has the main functionality associated with substrate. Moreover, the '-int_surf som' is only marginally (10-20%) slower than the '-int_surf img'.
Small comparison of simulation results of a sphere on the surface was performed to compare with (Loke&Menguc,2010) - Figs.10,11. Qualitatively the results are very similar, but there are certain quantitative differences. We will do a more careful benchmarking with the FFT version

Options '-surf <mre> <mim> <h>' was changed to '-surf <h> {<mre> <mim>|inf}'. So the order of parameters has been changed and option to specify perfectly reflecting substrate is now available. The results of '-surf ... inf' were checked to correspond to the limit of increasing <mre> to very large values. For perfectly reflecting substrate the default (and only) method to calculate reflected Green's tensor ('-int surf ...') is the image-dipole one.

Version incremented to 1.3a2. Fixed minor bug in sparse_ops.h, introduced recently. Added function sizetVector to memory.c/h. Call to imExp in CalcReflTerm_img was replaced by accImExp.