Scattnlay issue fixed
There were problems (which I already forgot about) with the fact that the Scattnlay and Bhfield programs give out an exact internal electric field with x-polarization. ADDA, by default, uses a y-polarized field. Therefore, I had to use the option "-orient 90 0 0" to rotate the coordinate system. In this case, simply changing the columns Ex and Ey (for real and imaginary parts) in the file with the exact field does not lead to the correct result (which is known from the paper).